Calculations for capability

+++ backend/src/main/java/com/bavostepbros/leap/domain/service/capabilityservice/CapabilityService.java
@@ -16,7 +16,7 @@ import com.bavostepbros.leap.domain.model.Resource;
 public interface CapabilityService {
 	Capability save(Integer environmentId, Integer statusId, Integer parentCapabilityId, String capabilityName,
 			String capabilityDescription, String paceOfChange, String targetOperatingModel, Integer resourceQuality,
-			Integer informationQuality, Integer applicationFit);
+			Double informationQuality, Double applicationFit);
 
 	Capability save (Capability capability);
 
@@ -26,7 +26,7 @@ public interface CapabilityService {
 
 	Capability update(Integer capabilityId, Integer environmentId, Integer statusId, Integer parentCapabilityId,
 			String capabilityName, String capabilityDescription, String paceOfChange, String targetOperatingModel,
-			Integer resourceQuality, Integer informationQuality, Integer applicationFit);
+			Integer resourceQuality, Double informationQuality, Double applicationFit);
 
 	void updateLevel(Capability capability);
